+++ packages/nextjs/app/participants/_components.tsx/ParticipantsTable.tsx
@@ -1,21 +1,20 @@
 "use client";
 
+import { SortableTable } from "./SortableTable";
 import { useScaffoldEventHistory } from "~~/hooks/scaffold-eth";
 import { calculateParticipantData } from "~~/utils/scaffold-eth";
-import { SortableTable } from "./SortableTable";
 
 export const ParticipantsTable = () => {
+  const { data: stakeEvents } = useScaffoldEventHistory({
+    contractName: "Staker",
+    eventName: "Stake",
+    fromBlock: 0n,
+  });
+  const participants = stakeEvents ? calculateParticipantData(stakeEvents) : [];
 
-    const { data: stakeEvents } = useScaffoldEventHistory({
-        contractName: "Staker",
-        eventName: "Stake",
-        fromBlock: 0n,
-    });
-    const participants = stakeEvents ? calculateParticipantData(stakeEvents) : [];
-
-    return (
-        <div className="overflow-x-auto">
-            <SortableTable data={participants} />
-        </div>
-    );
+  return (
+    <div className="overflow-x-auto" style={{ padding: "2rem" }}>
+      <SortableTable data={participants} />
+    </div>
+  );
 };
