+++ packages/nextjs/styles/globals.css
@@ -159,6 +159,14 @@ th {
   text-align: left;
 }
 
+.moreinfo {
+  cursor: pointer;
+  text-decoration: underline;
+  color: darkblue;
+  font-size: 0.8rem;
+  opacity: 0.8;
+}
+
 /* -- EXTENSION OVERRIDES -- */
 
 @plugin "daisyui/theme" {
